backend/python-vlv-service/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
Pillow
einops
accelerate
gliner
